:art: Create and style login page
