+++ src/WeedStore.js
@@ -1,19 +1,19 @@
-import { observable, action, computed, decorate } from "mobx"
+import { observable, action, computed, decorate } from "mobx";
 
 class WeedStore {
-    allStrains = [];
-    currentStrains = [];
-    
+  allStrains = [];
+  currentStrains = [];
+
 }
 
-decorate(
-    WeedStore, {
-        allStrains: observable,
-        currentStrains: observable,
-        // fetchStrains: action,
-        // setStrainsData: computed,
-    }
-)
+// decorate (
+//    WeedStore, {
+//      allStrains: observable,
+//      currentStrains: observable,
+//      // fetchStrains: action,
+//      // setStrainsData: computed,
+//    }
+// )
 
-const store = new WeedStore()
-export default store
+const store = new WeedStore();
+export default store;
